requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.28
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
cryptography==42.0.5
PyJWT==2.8.0
Werkzeug==3.0.1
requests==2.31.0
pandas==2.2.1
pytest==8.0.2
black==24.2.0
flake8==7.0.0
python-dateutil==2.8.2
gunicorn==21.2.0
bcrypt==4.1.2

